requirements.txt
streamlit
snowflake-connector-python
bcrypt
pandas
requests
streamlit-lottie
